Squashed commit of the following:

commit 12b93b4d6d02dc207f1d3cf6b01ed6df6c8e5a0a
Author: Jason Yin <jason.yin@oracle.com>
Date:   Tue Jul 10 19:59:53 2018 -0700

    int64 change

+++ filestorage/export_set.go
@@ -52,7 +52,7 @@ type ExportSet struct {
 	// `maxFsStatBytes` minus the metered size of the file
 	// system. If the metered size is larger than `maxFsStatBytes`,
 	// then `fbytes` and `abytes` will both be '0'.
-	MaxFsStatBytes *int `mandatory:"false" json:"maxFsStatBytes"`
+	MaxFsStatBytes *int64 `mandatory:"false" json:"maxFsStatBytes"`
 
 	// Controls the maximum `ffiles`, `ffiles`, and `afiles`
 	// values reported by `NFS FSSTAT` calls through any associated
@@ -63,7 +63,7 @@ type ExportSet struct {
 	// `maxFsStatFiles` minus the metered size of the file
 	// system. If the metered size is larger than `maxFsStatFiles`,
 	// then `ffiles` and `afiles` will both be '0'.
-	MaxFsStatFiles *int `mandatory:"false" json:"maxFsStatFiles"`
+	MaxFsStatFiles *int64 `mandatory:"false" json:"maxFsStatFiles"`
 }
 
 func (m ExportSet) String() string {
